refactor: agent 분석/추천 코드 분리
