+++ services/agent-service/web/src/app/(shell)/agent-result-shared.ts
@@ -1,0 +1,13 @@
+import type { BudgetForecastResponse } from "./analysis-service"
+import type { RecommendationQueryResponse } from "./recommendation-service"
+
+export type AnalysisResult = {
+  keyId: number
+  keyLabel: string
+  provider: string
+  data?: BudgetForecastResponse
+  recommendation?: RecommendationQueryResponse
+  error?: string
+  recommendationError?: string
+  forecastGaps?: string[]
+}
